+++ src/sokoenginepy/__init__.py
@@ -1,3 +1,3 @@
 from . import game, io
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"